Fix failing tests and extract platform.py
